requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
alembic==1.17.0
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
beautifulsoup4==4.14.2
    # via -r requirements.in
argon2-cffi==23.1.0
    # via
    #   -r requirements.in
    #   passlib
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
charset-normalizer==3.4.4
    # via requests
cffi==2.0.0
    # via
    #   argon2-cffi-bindings
    #   cryptography
click==8.3.0
    # via uvicorn
cryptography==46.0.3
    # via python-jose
ecdsa==0.19.1
    # via python-jose
fastapi==0.119.1
    # via -r requirements.in
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.27.2
    # via -r requirements.in
idna==3.11
    # via
    #   anyio
    #   httpx
joblib==1.5.2
    # via
    #   -r requirements.in
    #   scikit-learn
loguru==0.7.3
    # via -r requirements.in
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via mako
numpy==2.3.4
    # via
    #   -r requirements.in
    #   pandas
    #   scikit-learn
pandas==2.3.3
    # via -r requirements.in
pymupdf==1.24.10
    # via -r requirements.in
passlib==1.7.4
    # via -r requirements.in
psycopg2-binary==2.9.11
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycparser==2.23
    # via cffi
pydantic==2.12.3
    # via
    #   fastapi
    #   pydantic-settings
    #   sqlmodel
pydantic-core==2.41.4
    # via pydantic
pydantic-settings==2.11.0
    # via -r requirements.in
pyjwt==2.10.1
    # via redis
python-dotenv==1.1.1
    # via
    #   pydantic-settings
    #   uvicorn
python-dateutil==2.9.0.post0
    # via pandas
python-jose==3.5.0
    # via -r requirements.in
python-multipart==0.0.20
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via uvicorn
redis==5.3.1
    # via -r requirements.in
requests==2.32.5
    # via -r requirements.in
rsa==4.9.1
    # via python-jose
scikit-learn==1.7.2
    # via -r requirements.in
scipy==1.16.3
    # via
    #   -r requirements.in
    #   scikit-learn
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
soupsieve==2.8
    # via beautifulsoup4
sqlalchemy==2.0.44
    # via
    #   alembic
    #   sqlmodel
sqlmodel==0.0.21
    # via -r requirements.in
starlette==0.48.0
    # via fastapi
threadpoolctl==3.6.0
    # via
    #   -r requirements.in
    #   scikit-learn
typing-extensions==4.15.0
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
uvicorn==0.38.0
    # via -r requirements.in
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via uvicorn
